requirements.txt
# Core
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Optimization
cvxpy>=1.4.0
ortools>=9.7

# Bayesian
pymc>=5.10.0
arviz>=0.17.0

# Time Series
statsmodels>=0.14.0
arch>=6.2.0

# Data Access
fredapi>=0.5.1
requests>=2.31.0
wbgapi>=1.0.12

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.27.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
